requirements.txt
torch
transformers
pandas
numpy
scikit-learn
biopython
obonet
networkx
tqdm
pyyaml
